+++ bindings/nostr-sdk-js/src/client/mod.rs
@@ -60,12 +60,6 @@ impl JsClient {
         JsClientBuilder::new()
     }
 
-    /// Update default difficulty for new `Event`
-    #[wasm_bindgen(js_name = updateDifficulty)]
-    pub fn update_difficulty(&self, difficulty: u8) {
-        self.inner.update_difficulty(difficulty);
-    }
-
     /// Update minimum POW difficulty for received events
     ///
     /// Events with a POW lower than the current value will be ignored to prevent resources exhaustion.